+++ EasyAccept.Core/Interpreter/Commands/EchoCommand.cs
@@ -1,22 +1,14 @@
 using EasyAccept.Core.Interpreter.Arguments;
-using EasyAccept.Core.Output;
+using EasyAccept.Core.Interpreter.Results;
 
 namespace EasyAccept.Core.Interpreter.Commands
 {
   public class EchoCommand : ICommand
   {
     private readonly NonNamedArgument Argument;
-    private readonly IOutputDriver OutputDriver;
 
-    public EchoCommand(NonNamedArgument argument, IOutputDriver outputDriver)
-    {
-      Argument = argument;
-      OutputDriver = outputDriver;
-    }
+    public EchoCommand(NonNamedArgument argument) => Argument = argument;
 
-    public void Execute()
-    {
-      OutputDriver.WriteLine(Argument.ToString());
-    }
+    public IResult Execute() => new PrintableResult(Argument.ToString());
   }
 }