+++ EasyAccept.Core/Interpreter/Commands/ICommand.cs
@@ -1,7 +1,9 @@
+using EasyAccept.Core.Interpreter.Results;
+
 namespace EasyAccept.Core.Interpreter.Commands
 {
   public interface ICommand
   {
-    void Execute();
+    IResult Execute();
   }
 }