+++ EasyAccept.Core/Interpreter/Exceptions/SyntaxException.cs
@@ -1,9 +1,0 @@
-using EasyAccept.Core.Exceptions;
-
-namespace EasyAccept.Core.Interpreter.Exceptions
-{
-  public class SyntaxException : EasyAcceptException
-  {
-    public SyntaxException(string message) : base(message) { }
-  }
-}