+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.1.0](https://github.com/EasyAccept/easyaccept-csharp/releases/tag/1.1.0)
+
+- GetCompleteResults refactored to return a more equivalent return as the original EasyAccept
+- New methods on EasyAcceptFacade added, as GetSummarizedResults, GetTotalNumberOfPassedTests, GetTotalNumberOfTests, GetScriptFailures and others
+
 ## [1.0.4](https://github.com/EasyAccept/easyaccept-csharp/releases/tag/1.0.4)
 
 - First published version on NuGet